move config files to a seperate package to clean the code a bit
