+++ confutils/go.mod
@@ -1,0 +1,7 @@
+module confutils
+
+go 1.16
+
+require (
+	github.com/pelletier/go-toml/v2 v2.0.0-beta.4
+)
