+++ go.mod
@@ -3,8 +3,9 @@ module rounds
 go 1.16
 
 require (
-	github.com/golang/protobuf v1.5.2 // indirect
+	confutils v1.0.0
 	github.com/hajimehoshi/ebiten/v2 v2.2.1
-	github.com/pelletier/go-toml/v2 v2.0.0-beta.4 // indirect
 	google.golang.org/protobuf v1.27.1
 )
+
+replace confutils v1.0.0 => ./confutils
